refactor: folder structure refactor
